feat: 2D renderer
BUG: forgot about the rendering components' IDs....

+++ DinnEngine/src/DinnCore/Rendering/Components/VertexBufferObject.cpp
@@ -4,7 +4,7 @@
 
 namespace Dinn 
 {
-	VBO::VertexBufferObject(uint32_t* vertices, std::size_t size)
+	VBO::VertexBufferObject(float* vertices, std::size_t size)
 	{
 		glGenBuffers(1, &ID);
 		glBindBuffer(GL_ARRAY_BUFFER, ID);
